Added colors for erc-colorize
erc-colorize is a mode that assign a color to each user in an IRC chat
room with a unique color, so text written by user a is in red and text
written by user b is blue, etc, making it easy to follow a conversation
in an ERC buffer.

I reused some of the colors defined by the rainbow-delimiters. They look
really nice on the dark version, and should look good on the light
version.